Initial Scaffolding of Data Tables - TrackTour API
This commit is the initial rails api scaffolding of the various data
tables that will be used in the TrackTour web application.

Here are the following data tables that were created.

Artists - This is the artist table that will contain a list of artists
that the user can put on their 'wish list' to see.

Wishlists - This is a bridge table between users and the artists they
want to see in concert.  This is the highest level of collection
organization used in this application.

Concerts - This is a list of concerts where multiple artists may have
performed. This will be used by users to track where they have seen
artists on their list.

Attended Concerts - This is a bridge table between the 'wishlist' or the
artist that a user wants to see in concert and the concert where a user
saw them.  This bridge table allows us to accommodate users seeing an
artist perform on more than one occasion and allows for future
aggregation.